fix: move tests up a folder
