+++ packages/dynamic/test/dynamicsvgexample.js
@@ -1,4 +1,4 @@
-const helper = require("../helpers/truffleTestHelper");
+const helper = require("./helpers/truffleTestHelper");
 const truffleAssert = require('truffle-assertions');
 const ERC721Creator = artifacts.require('MockERC721Creator');
 const DynamicSVGExample = artifacts.require("DynamicSVGExample");
