+++ packages/manifold/test/burnredeem1155.js
