+++ packages/manifold/test/creatoroperatorfilter.js
