+++ packages/manifold/test/operatorfilterer.js
